updates [skip ci]

+++ dserver/entrypoint.sh
@@ -1,13 +1,14 @@
 #!/bin/bash
-set -e
+set -xe
 #
-cd "${HOME}"
-git clone --recursive https://github.com/forwardcomputers/dfiles.git tmp
-mv tmp/.git .
-rm -rf tmp
-rm -rf tmp/ .profile
-git reset --hard
-git remote set-url origin ssh://git@github.com/forwardcomputers/dfiles.git
+if [[ ! -d "${HOME}"/.git ]]; then
+    cd "${HOME}"
+    git clone --recursive https://github.com/forwardcomputers/dfiles.git tmp
+    mv tmp/.git .
+    rm -rf tmp
+    rm -rf tmp/ .profile
+    git reset --hard
+    git remote set-url origin ssh://git@github.com/forwardcomputers/dfiles.git
+fi
 #
-#sudo bash -c 'cp /etc/hosts /tmp/hosts.new && sed "s/^127.0.0.1.*/127.0.0.1\tlocalhost "$(cat /etc/hostname)"/" -i /tmp/hosts.new && cp -f /tmp/hosts.new /etc/hosts'
-sudo bash -c '/usr/sbin/sshd -D'
+sudo bash -c '/usr/sbin/sshd -Ded'
